init each target stack as ivar; re-use as required
This change allows us to re-use the same @stack ivar across multiple operations,
like those in meta-subcommands like `plan-all`. Rather than re-init'd the stack
object each time, we re-use the same object. The benefit of this is that we can
preserve and present some initial information when the stack is first init'd
and not repeat ourselves over & over for every operation. However, because this
requires us to unset the ivar at the end of a meta-subcommand, it is sub-optimal
and error-prone (i.e. `@stack = nil`). This needs to be refactored.