fix broken upgrade routines
* make `upgrade` a meta-subcommand:
  * add `clean`, `init` and `fmt` to the list of chores when upgrading a stack
  * breakup monolithic method
  * provide better feedback
  * fix malformed `required_version` expression
  * fix `required_version` regex catch-22

The first rendition of this feature was broken in TWO ways ... First, it couldn't parse the `required_version` form the HCL because the order of operations for setting the replacement (backref) wasn't always evaluated in time. Weird.

Secondly, there was a bug in the text replacement routine that malformed the `required_version` expression, by adding a second curly brace.

All this is fixed and we have better feedback and handling.